#446 Refacto on token/jwk handlers, the authentication of a incoming/outcoming request isn't bound anymore with the token/keys generation
