+++ bot/connector-teams/src/main/kotlin/TeamsConnector.kt
@@ -25,8 +25,9 @@ import fr.vsct.tock.bot.connector.ConnectorCallback
 import fr.vsct.tock.bot.connector.ConnectorData
 import fr.vsct.tock.bot.connector.teams.auth.AuthenticateBotConnectorService
 import fr.vsct.tock.bot.connector.teams.auth.ForbiddenException
+import fr.vsct.tock.bot.connector.teams.auth.JWKHandler.launchJWKCollector
 import fr.vsct.tock.bot.connector.teams.messages.SendActionConverter
-import fr.vsct.tock.bot.connector.teams.token.TokenHandler.takeCareOfTheToken
+import fr.vsct.tock.bot.connector.teams.token.TokenHandler.launchTokenCollector
 import fr.vsct.tock.bot.engine.BotRepository
 import fr.vsct.tock.bot.engine.ConnectorController
 import fr.vsct.tock.bot.engine.action.SendSentence
@@ -61,7 +62,8 @@ internal class TeamsConnector(
     private val authenticateBotConnectorService = AuthenticateBotConnectorService(appId)
 
     override fun register(controller: ConnectorController) {
-        takeCareOfTheToken(appId, appPassword)
+        launchTokenCollector(appId, appPassword)
+        launchJWKCollector()
 
         controller.registerServices(path) { router ->
